requirements.txt
transformers==4.50.0
qwen-vl-utils[decord]==0.0.10
torch==2.6.0
torchvision
torchaudio
json-repair
numpy==2.0.2
Pillow==10.3.0
Requests==2.31.0
datasets
omegaconf
tiktoken>=0.7.0
addict
yapf
timm
opencv-python
supervision
pycocotools
accelerate
sentencepiece
spacy
peft
super-image
einops
protobuf
deepspeed
librosa==0.9.0
soundfile==0.12.1
vector-quantize-pytorch==1.18.5
vocos==0.1.0